refactor: 썸네일 그리는 로직 커스텀 훅으로 분리
