+++ src/components/@shared/Thumbnail.tsx
@@ -1,43 +1,11 @@
-import { useEffect, useRef, useState } from "react";
 import styled from "styled-components";
-import {
-  RGBA_ARRAY_SIZE,
-  START_X,
-  START_Y,
-  THUMBNAIL_HEIGHT,
-  THUMBNAIL_WIDTH,
-  TRIANGLE_COUNT,
-} from "../../constants";
+import { THUMBNAIL_HEIGHT, THUMBNAIL_WIDTH } from "../../constants";
+import useDrawThumbnail from "../../hooks/useDrawThumbnail";
 import { getCoordinate } from "../../utils/utils";
 import Triangle from "./Triangle";
 
 const Thumbnail = ({ src }: ThumbnailProps) => {
-  const triangleArray = Array.from({ length: TRIANGLE_COUNT }, () => 0);
-  const canvasRef = useRef<HTMLCanvasElement>(null);
-  const imageRef = useRef<HTMLImageElement>(null);
-  const [rgba, setRgba] = useState<Uint8ClampedArray[] | null>(null);
-
-  useEffect(() => {
-    if (canvasRef.current && imageRef.current) {
-      const canvas = canvasRef.current.getContext("2d", { willReadFrequently: true });
-      canvas?.drawImage(imageRef.current, START_X, START_Y, THUMBNAIL_WIDTH, THUMBNAIL_HEIGHT);
-
-      if (canvas?.getImageData(START_X, START_Y, THUMBNAIL_WIDTH, THUMBNAIL_HEIGHT)) {
-        const rgbaArray = [];
-        const rgbaData = canvas?.getImageData(
-          START_X,
-          START_Y,
-          THUMBNAIL_WIDTH,
-          THUMBNAIL_HEIGHT
-        ).data;
-
-        for (let i = 0; i < rgbaData.length; i += RGBA_ARRAY_SIZE) {
-          rgbaArray.push(rgbaData.slice(i, i + RGBA_ARRAY_SIZE));
-        }
-        setRgba(rgbaArray);
-      }
-    }
-  }, [imageRef.current]);
+  const { triangleArray, canvasRef, imageRef, rgba } = useDrawThumbnail();
 
   return (
     <S.Container>
@@ -80,6 +48,7 @@ const S = {
   Canvas: styled.canvas`
     width: inherit;
     height: inherit;
+    filter: blur(5px);
   `,
 
   Image: styled.img`
